% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/balanced_with_interaction.R
\name{balanced_with_interaction}
\alias{balanced_with_interaction}
\title{Balanced With Interaction Analysis}
\usage{
balanced_with_interaction(
  data,
  part = P,
  operator = O,
  measurement = Y,
  alpha = 0.05,
  conf_type = "mls"
)
}
\arguments{
\item{data}{a data frame that contains measurements, operators and parts
for a Gauge R&R analysis}

\item{part}{a column of the data frame that has the part labels for the
measurements}

\item{operator}{a column of the data frame that has the operator labels
for the measurements}

\item{measurement}{a column of the data frame that has measurements of the
object collected}

\item{alpha}{the value for the confidence interval calculation (i.e. 95\% CI
would have alpha=0.05)}

\item{conf_type}{specifying the type of confidence interval wanted to calculation,
and there are two options ("mls" or "gpq")}
}
\value{
returns a data frame with the confidence intervals of the data that is
passed into the function
}
\description{
Balanced With Interaction Analysis
}
\examples{
mydata <- data.frame(P=c(1,1,1,2,2,2), O=c(1,2,3,1,2,3), Y=c(2,2.1,2.2,1.9,2.3,1.8))
balanced_with_interaction(mydata, part=P, operator=O, measurement=Y, alpha=0.01)
balanced_with_interaction(mydata, conf_type = "gpq")
}
