% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conf_intervals.R
\name{conf_intervals}
\alias{conf_intervals}
\title{Modified large-sample (MLS) Confidence Interval Calculation}
\usage{
conf_intervals(data, part = P, operator = O, measurement = Y, alpha = 0.05)
}
\arguments{
\item{data}{a data frame that contains measurements, operators and parts
for a Gauge R&R analysis}

\item{part}{a column of the data frame that has the part labels for the
measurements}

\item{operator}{a column of the data frame that has the operator labels
for the measurements}

\item{measurement}{a column of the data frame that has measurements of the
object collected}

\item{alpha}{the value for the confidence interval calculation (i.e. 95\% CI
would have alpha=0.05)}
}
\value{
a data frame with values of the point estimate and the upper and lower
bounds of the mls confidence interval.
}
\description{
Modified large-sample (MLS) Confidence Interval Calculation
}
\examples{
mydata <- data.frame(P=c(1,1,2,2,3,3),O=c(1,2,1,2,1,2),Y=c(5.3, 6.5, 5.4, 6.4, 6.9, 5.8))
conf_intervals(mydata, alpha=0.1)
conf_intervals(mydata, part=P, operator=O, measurement=Y, alpha=0.01)
}
