% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conf_intervals_gpq.R
\name{conf_intervals_gpq}
\alias{conf_intervals_gpq}
\title{Generalized Pivotal Quantities (GPQ) Confidence Intervals}
\usage{
conf_intervals_gpq(
  data,
  part = P,
  operator = O,
  measurement = Y,
  alpha = 0.05,
  N = 1e+05
)
}
\arguments{
\item{data}{a data frame that contains measurements, operators and parts
for a Gauge R&R analysis}

\item{part}{a column of the data frame that has the part labels for the
measurements}

\item{operator}{a column of the data frame that has the operator labels
for the measurements}

\item{measurement}{a column of the data frame that has measurements of the
object collected}

\item{alpha}{the value for the confidence interval calculation (i.e. 95\% CI
would have alpha=0.05)}

\item{N}{the number of simulations run for the CI calculation.}
}
\value{
a data frame with the value of the estimate and the upper and lower
bounds of the CI.
}
\description{
Generalized Pivotal Quantities (GPQ) Confidence Intervals
}
\examples{
mydata <- data.frame(P=c(2,2,4,4,5,5),O=c(4,4,4,4,4,4),Y=c(5.3, 6.5, 5.4, 6.4, 6.9, 5.8))
conf_intervals_gpq(mydata, alpha = 0.1)
conf_intervals_gpq(mydata, N=1000)
conf_intervals_gpq(mydata, part=P, operator=O, measurement=Y, alpha=0.01, N=1e4)
}
