% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gauge_variance.R
\name{gauge_variance}
\alias{gauge_variance}
\title{Variance Calculations for Gauge R&R data}
\usage{
gauge_variance(data, part = P, operator = O, measurement = Y)
}
\arguments{
\item{data}{a data frame that contains measurements, operators and parts
for a Gauge R&R analysis}

\item{part}{a column of the data frame that has the part labels for the
measurements (this is can be specified by just the column name i.e. P)}

\item{operator}{a column of the data frame that has the operator labels
for the measurements (this is can be specified by just the column name i.e. O)}

\item{measurement}{a column of the data frame that has measurements of the
object collected (this is can be specified by just the column name i.e. Y)}
}
\value{
a data frame with the variance values and the names of the quantities
that are associated with such values.
}
\description{
Variance Calculations for Gauge R&R data
}
\examples{
mydata <- data.frame(P=c(2,2,4,4,5,5),O=c(4,4,4,4,4,4),Y=c(5.3, 6.5, 5.4, 6.4, 6.9, 5.8))
gauge_variance(mydata)


}
