% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gauge_variance_per.R
\name{gauge_variance_per}
\alias{gauge_variance_per}
\title{Percent of Variance Contribution}
\usage{
gauge_variance_per(data, part = P, operator = O, measurement = Y)
}
\arguments{
\item{data}{a data frame that contains measurements, operators and parts
for a Gauge R&R analysis}

\item{part}{a column of the data frame that has the part labels for the
measurements}

\item{operator}{a column of the data frame that has the operator labels
for the measurements}

\item{measurement}{a column of the data frame that has measurements of the
object collected}
}
\value{
returns a data frame with the values of the percent of variance contribution
}
\description{
Percent of Variance Contribution
}
\examples{
mydata <- data.frame(p=c(1,1,2,2), o=c(1,2,1,2), m=c(2.3,4.5,3.4,4.3))
gauge_variance_per(mydata, part=p, operator=o, measurement=m)

}
