% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mean_ss.R
\name{mean_ss}
\alias{mean_ss}
\title{Mean Sum of Squares}
\usage{
mean_ss(data, part = P, operator = O, measurement = Y)
}
\arguments{
\item{data}{a data frame that contains measurements, operators and parts
for a Gauge R&R analysis}

\item{part}{a column of the data frame that has the part labels for the
measurements}

\item{operator}{a column of the data frame that has the operator labels
for the measurements}

\item{measurement}{a column of the data frame that has measurements of the
object collected}
}
\value{
The 4 mean sum of square values
MSE: which considers both part and operator
MSO: which just considers the operator
MSP: which just considers the part
MSPO: which considers the values for part, operator and part/operator interaction
}
\description{
Mean Sum of Squares
}
\examples{
mydata <- data.frame(P=c(2,2,4,4,5,5),O=c(4,4,4,4,4,4),Y=c(5.3, 6.5, 5.4, 6.4, 6.9, 5.8))
mean_ss(mydata, part=P, operator=O, measurement = Y)

}
