% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.intervals_table}
\alias{plot.intervals_table}
\title{Plotting Function}
\usage{
\method{plot}{intervals_table}(df, ...)
}
\arguments{
\item{df}{is a data frame object with class type "intervals_table"}

\item{...}{other arguments that can be passed into the function}
}
\value{
a ggplot of the confidence intervals defined by the table.
}
\description{
Plotting Function
}
\examples{
mydf <- structure(data.frame(quantity = c("part", "operator"),
estimate = c(12, 15), lower = c(3, 5), upper = c(20, 18)), class= "intervals_table")
mydf.plot
}
