% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unbalanced_with_interaction.R
\name{unbalanced_with_interaction}
\alias{unbalanced_with_interaction}
\title{Unbalanced Two Factor Model With Interaction}
\usage{
unbalanced_with_interaction(
  data,
  part = P,
  operator = O,
  measurement = Y,
  alpha = 0.05
)
}
\arguments{
\item{data}{a data frame that contains measurements, operators and parts
for a Gauge R&R analysis}

\item{part}{a column of the data frame that has the part labels for the
measurements (this is can be specified by just the column name i.e. P)}

\item{operator}{a column of the data frame that has the operator labels
for the measurements (this is can be specified by just the column name i.e. O)}

\item{measurement}{a column of the data frame that has measurements of the
object collected (this is can be specified by just the column name i.e. Y)}

\item{alpha}{the value for the confidence interval calculation (i.e. 95\% CI
would have alpha=0.05)}
}
\value{
a data frame of the values of the point estimates and the upper and
lower bounds for each estimate
}
\description{
Unbalanced Two Factor Model With Interaction
}
\examples{
mydata <- data.frame(P=c(1,1,2,2,3,3),O=c(1,2,1,2,1,2),Y=c(5.3, 6.5, 5.4, 6.4, 6.9, 5.8))
unbalanced_with_interaction(mydata, alpha=0.01)

}
\references{
Burdick, Richard K., Connie M. Borror, and Douglas C. Montgomery. Design and Analysis of Gauge R&R Studies: Making Decisions with Confidence Intervals in Random and Mixed ANOVA Models. Society for Industrial and Applied Mathematics, 2005.
}
