% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/balanced_no_interaction.R
\name{balanced_no_interaction}
\alias{balanced_no_interaction}
\title{Balanced Without Interaction}
\usage{
balanced_no_interaction(
  data,
  part = P,
  operator = O,
  measurement = Y,
  alpha = 0.05
)
}
\arguments{
\item{data}{a data frame that contains measurements, operators and parts
for a Gauge R&R analysis}

\item{part}{a column of the data frame that has the part labels for the
measurements}

\item{operator}{a column of the data frame that has the operator labels
for the measurements}

\item{measurement}{a column of the data frame that has measurements of the
object collected}

\item{alpha}{the value for the confidence interval calculation (i.e. 95\% CI
would have alpha=0.05)}
}
\value{
a data frame of the values of the point estimates and the upper and
lower bounds for each estimate
}
\description{
Balanced Without Interaction
}
