% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gauge_rr.R
\name{gauge_rr}
\alias{gauge_rr}
\title{General Gauge R&R Analysis Function}
\usage{
gauge_rr(data, part = P, operator = O, measurement = Y)
}
\arguments{
\item{data}{a data frame that contains measurements, operators and parts
for a Gauge R&R analysis}

\item{part}{a column of the data frame that has the part labels for the
measurements}

\item{operator}{a column of the data frame that has the operator labels
for the measurements}

\item{measurement}{a column of the data frame that has measurements of the
object collected}
}
\value{
a data frame with the confidence and point estimator values, for the
appropriate analysis (balanced/unbalanced and one factor/two factor.)
}
\description{
General Gauge R&R Analysis Function
}
