% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gauge_rr.R
\name{gauge_rr}
\alias{gauge_rr}
\title{General Gauge R&R Analysis Function}
\usage{
gauge_rr(
  data,
  part = P,
  operator = NULL,
  measurement = Y,
  alpha = 0.05,
  interaction = FALSE,
  formula = NULL
)
}
\arguments{
\item{data}{a data frame that contains measurements, operators and parts
for a Gauge R&R analysis}

\item{part}{a column of the data frame that has the part labels for the
measurements (this is can be specified by just the column name i.e. P)}

\item{operator}{a column of the data frame that has the operator labels
for the measurements (this is can be specified by just the column name i.e. O)}

\item{measurement}{a column of the data frame that has measurements of the
object collected (this is can be specified by just the column name i.e. Y)}

\item{alpha}{the value for the confidence interval calculation (i.e. 95\% CI
would have alpha=0.05)}

\item{interaction}{this is a binary parameter where the user can specify if
there is or is not interaction in the data set. (using T/F or TRUE/FALSE)}

\item{formula}{the user is able to input a formula that determines how many
factors the model has. This formula must have the lme4 format (i.e. y ~ (1 | P) + (1 | O))}
}
\value{
a data frame with the confidence and point estimator values, for the
appropriate analysis (balanced/unbalanced, one factor/two factor, and
interaction/no interaction)
}
\description{
General Gauge R&R Analysis Function
}
\examples{
mydata <- data.frame(P=c(1,2,3,4,1,2,3,4), Y=c(2,2,2,3,3,2,3,2))
gauge_rr(mydata, part = P, measurement = Y, interaction=FALSE)

}
\references{
Burdick, Richard K., Connie M. Borror, and Douglas C. Montgomery. Design and Analysis of Gauge R&R Studies: Making Decisions with Confidence Intervals in Random and Mixed ANOVA Models. Society for Industrial and Applied Mathematics, 2005.
}
